rejig the type imports to avoid circular deps
