+++ packages/react-search-ui-views/package.json
@@ -4,7 +4,7 @@
   "description": "A collection of React UI components for building search experiences",
   "license": "Apache-2.0",
   "main": "lib/cjs",
-  "module": "lib/es",
+  "module": "lib/esm",
   "sideEffects": [
     "*.css"
   ],
@@ -71,7 +71,6 @@
     "downshift": "^3.2.10",
     "rc-pagination": "^1.20.1",
     "react-select": "^2.4.4",
-    "@elastic/search-ui": "^1.9.0",
-    "@elastic/react-search-ui": "^1.9.0"
+    "@elastic/search-ui": "^1.9.0"
   }
 }
