+++ packages/react-search-ui-views/src/PagingInfo.tsx
@@ -1,7 +1,28 @@
-import { PagingInfoViewProps } from "@elastic/react-search-ui";
+import { SearchContextState } from "@elastic/search-ui";
 import React from "react";
+import { Rename, BaseContainerProps } from "./types";
 
 import { appendClassName } from "./view-helpers";
+
+export type PagingInfoContainerContext = Pick<
+  SearchContextState,
+  "pagingStart" | "pagingEnd" | "resultSearchTerm" | "totalResults"
+>;
+
+export type PagingInfoViewProps = Rename<
+  BaseContainerProps & PagingInfoContainerContext,
+  {
+    pagingStart: "start";
+    resultSearchTerm: "searchTerm";
+    pagingEnd: "end";
+  }
+>;
+
+export type PagingInfoContainerProps = BaseContainerProps &
+  PagingInfoContainerContext & {
+    view?: React.ComponentType<PagingInfoViewProps>;
+  };
+
 
 function PagingInfo({
   className,
