+++ packages/react-search-ui-views/src/__tests__/BooleanFacet.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import BooleanFacet from "../BooleanFacet";
 import { shallow, mount } from "enzyme";
-import { FacetViewProps } from "@elastic/react-search-ui";
+import { FacetViewProps } from "../Facet";
 
 const params: FacetViewProps = {
   label: "A Facet",
