+++ packages/react-search-ui/src/containers/ErrorBoundary.tsx
@@ -1,13 +1,11 @@
 import React, { Component } from "react";
 import { withSearch } from "..";
-import { ErrorBoundary } from "@elastic/react-search-ui-views";
-import { SearchContextState } from "../withSearch";
+import {
+  ErrorBoundary,
+  ErrorBoundaryContainerContext,
+  ErrorBoundaryViewProps
+} from "@elastic/react-search-ui-views";
 import { BaseContainerProps } from "../types";
-
-type ErrorBoundaryContainerContext = Pick<SearchContextState, "error">;
-
-export type ErrorBoundaryViewProps = BaseContainerProps &
-  ErrorBoundaryContainerContext;
 
 type ErrorBoundaryContainerProps = BaseContainerProps &
   ErrorBoundaryContainerContext & {
