+++ packages/react-search-ui/src/containers/Facet.tsx
@@ -1,53 +1,17 @@
 import React from "react";
 import { Component } from "react";
-import { MultiCheckboxFacet } from "@elastic/react-search-ui-views";
 import {
-  helpers,
-  FilterType,
-  FacetValue,
-  FieldValue
-} from "@elastic/search-ui";
+  FacetContainerContext,
+  FacetContainerProps,
+  FacetViewProps,
+  MultiCheckboxFacet
+} from "@elastic/react-search-ui-views";
+import { helpers } from "@elastic/search-ui";
 
 import { accentFold } from "../helpers";
 import { withSearch } from "..";
-import { BaseContainerProps } from "../types";
-import { SearchContextState } from "../withSearch";
 
 const { markSelectedFacetValuesFromFilters } = helpers;
-
-export type FacetViewProps = {
-  className?: string;
-  label: string;
-  onMoreClick: () => void;
-  onRemove: (value: FieldValue) => void;
-  onChange: (value: FieldValue) => void;
-  onSelect: (value: FieldValue) => void;
-  options: FacetValue[];
-  showMore: boolean;
-  values: any;
-  showSearch: boolean;
-  onSearch: (value: string) => void;
-  searchPlaceholder: string;
-};
-
-type FacetContainerProps = BaseContainerProps & {
-  filterType: FilterType;
-  show?: number;
-  view?: React.ComponentType<FacetViewProps>;
-  isFilterable: boolean;
-  field: string;
-  label: string;
-};
-
-type FacetContainerContextProps = Pick<
-  SearchContextState,
-  | "addFilter"
-  | "removeFilter"
-  | "setFilter"
-  | "a11yNotify"
-  | "filters"
-  | "facets"
->;
 
 type FacetContainerState = {
   searchTerm: string;
@@ -55,7 +19,7 @@ type FacetContainerState = {
 };
 
 export class FacetContainer extends Component<
-  FacetContainerProps & FacetContainerContextProps,
+  FacetContainerProps,
   FacetContainerState
 > {
   static defaultProps = {
@@ -166,7 +130,7 @@ export class FacetContainer extends Component<
   }
 }
 
-export default withSearch<FacetContainerProps, FacetContainerContextProps>(
+export default withSearch<FacetContainerProps, FacetContainerContext>(
   ({ filters, facets, addFilter, removeFilter, setFilter, a11yNotify }) => ({
     filters,
     facets,
