+++ packages/react-search-ui/src/containers/Results.tsx
@@ -1,33 +1,18 @@
 import React, { Component } from "react";
-import { Result, Results } from "@elastic/react-search-ui-views";
+import {
+  Result,
+  Results,
+  ResultsContainerProps,
+  ResultsContainerContext
+} from "@elastic/react-search-ui-views";
 
 import { withSearch } from "..";
 import { Result as ResultContainer } from ".";
-import { BaseContainerProps } from "../types";
-import { SearchContextState } from "../withSearch";
-import { ResultViewProps } from "./Result";
-import { SearchResult } from "@elastic/search-ui";
 
 function getRaw(result, value) {
   if (!result[value] || !result[value].raw) return;
   return result[value].raw;
 }
-
-type ResultsContainerContext = Pick<SearchContextState, "results">;
-
-type ResultsContainerProps = BaseContainerProps &
-  ResultsContainerContext & {
-    view?: React.ComponentType<ResultsViewProps>;
-    resultView?: React.ComponentType<ResultViewProps>;
-    clickThroughTags?: string[];
-    titleField?: string;
-    urlField?: string;
-    thumbnailField?: string;
-    results: SearchResult[];
-    shouldTrackClickThrough?: boolean;
-  };
-
-export type ResultsViewProps = BaseContainerProps
 
 export class ResultsContainer extends Component<ResultsContainerProps> {
   static defaultProps = {
