+++ packages/react-search-ui/src/containers/ResultsPerPage.tsx
@@ -1,29 +1,11 @@
 import React, { Component } from "react";
 import { withSearch } from "..";
-import { ResultsPerPage } from "@elastic/react-search-ui-views";
-import { SearchContextState } from "../withSearch";
-import { BaseContainerProps, Rename } from "../types";
-
-type ResultsPerPageContainerContext = Pick<
-  SearchContextState,
-  "resultsPerPage" | "setResultsPerPage"
->;
-
-type ResultsPerPageContainerProps = BaseContainerProps &
-  ResultsPerPageContainerContext & {
-    view?: React.ComponentType<ResultsPerPageViewProps>;
-    options?: number[];
-  };
-
-export type ResultsPerPageViewProps = BaseContainerProps &
-  Pick<ResultsPerPageContainerProps, "options"> &
-  Rename<
-    ResultsPerPageContainerContext,
-    {
-      setResultsPerPage: "onChange";
-      resultsPerPage: "value";
-    }
-  >;
+import {
+  ResultsPerPage,
+  ResultsPerPageContainerContext,
+  ResultsPerPageContainerProps,
+  ResultsPerPageViewProps
+} from "@elastic/react-search-ui-views";
 
 export class ResultsPerPageContainer extends Component<ResultsPerPageContainerProps> {
   static defaultProps = {
