+++ packages/react-search-ui/src/containers/Sorting.tsx
@@ -1,9 +1,11 @@
 import React, { Component } from "react";
 import { withSearch } from "..";
-import { Sorting } from "@elastic/react-search-ui-views";
-
-import { BaseContainerProps } from "../types";
-import { SearchContextState } from "../withSearch";
+import {
+  Sorting,
+  SortingContainerContext,
+  SortingContainerProps,
+  SortingViewProps
+} from "@elastic/react-search-ui-views";
 
 function findSortOption(sortOptions, sortData) {
   if (sortData.indexOf("|||") === -1) {
@@ -41,28 +43,6 @@ function formatSelectOption(sortOption) {
     value: formatSelectValues(sortOption)
   };
 }
-
-type SortingContainerContext = Pick<
-  SearchContextState,
-  "sortDirection" | "sortField" | "sortList" | "setSort"
->;
-
-export type SortingViewProps = BaseContainerProps &
-  Pick<SortingContainerProps, "label"> & {
-    onChange: (sortData?: any) => void;
-    options: {
-      value: string;
-      label: string;
-    }[];
-    value: string;
-  };
-
-type SortingContainerProps = BaseContainerProps &
-  SortingContainerContext & {
-    view?: React.ComponentType<SortingViewProps>;
-    label?: string;
-    sortOptions: any;
-  };
 
 export class SortingContainer extends Component<SortingContainerProps> {
   render() {
