+++ packages/react-search-ui/src/containers/__tests__/Facet.test.tsx
@@ -1,7 +1,8 @@
 import React from "react";
-import { FacetContainer, FacetViewProps } from "../Facet";
+import { FacetContainer } from "../Facet";
 import { shallow, ShallowWrapper } from "enzyme";
 import { Filter } from "@elastic/search-ui";
+import { FacetViewProps } from "@elastic/react-search-ui-views";
 
 const View = () => <div />;
