+++ packages/react-search-ui/src/containers/__tests__/Result.test.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import { shallow } from "enzyme";
-import { ResultContainer, ResultViewProps } from "../Result";
+import { ResultContainer } from "../Result";
+import { ResultViewProps } from "@elastic/react-search-ui-views";
 
 const params = {
   result: {
