+++ packages/react-search-ui/src/containers/index.ts
@@ -7,17 +7,3 @@ export { default as Results } from "./Results";
 export { default as ResultsPerPage } from "./ResultsPerPage";
 export { default as SearchBox } from "./SearchBox";
 export { default as Sorting } from "./Sorting";
-
-export type { FacetViewProps } from "./Facet";
-export type { PagingViewProps } from "./Paging";
-export type { PagingInfoViewProps } from "./PagingInfo";
-export type { ResultViewProps } from "./Result";
-export type { ResultsViewProps } from "./Results";
-export type { ResultsPerPageViewProps } from "./ResultsPerPage";
-export type {
-  SearchBoxViewProps,
-  SearchBoxAutocompleteViewProps,
-  InputViewProps
-} from "./SearchBox";
-export type { SortingViewProps } from "./Sorting";
-export type { ErrorBoundaryViewProps } from "./ErrorBoundary";
